log domain events (i.e. adding/changing/deleting entities)

+++ src/service/DeviceService.kt
@@ -1,29 +1,43 @@
 package cloud.fabx.service
 
+import cloud.fabx.application.AuthorizationException
+import cloud.fabx.application.DevicePrincipal
+import cloud.fabx.application.NewDevicePrincipal
+import cloud.fabx.application.XPrincipal
 import cloud.fabx.db.DbHandler.dbQuery
+import cloud.fabx.domainEvent
 import cloud.fabx.dto.DeviceDto
 import cloud.fabx.dto.EditDeviceDto
 import cloud.fabx.dto.NewDeviceDto
+import cloud.fabx.logger
 import cloud.fabx.model.Device
 import cloud.fabx.model.Devices
+import net.logstash.logback.argument.StructuredArguments.keyValue
 
 class DeviceService {
 
+    private val log = logger()
+
     private val toolService = ToolService()
 
-    suspend fun getAllDevices(): List<DeviceDto> = dbQuery {
+    suspend fun getAllDevices(principal: XPrincipal): List<DeviceDto> = dbQuery {
+        principal.requirePermission("get all devices", XPrincipal::allowedToGetAllDevices)
         Device.all().map{ toDeviceDto(it) }.toCollection(ArrayList())
     }
 
-    suspend fun getDeviceById(id: Int): DeviceDto? = dbQuery {
+    suspend fun getDeviceById(id: Int, principal: XPrincipal): DeviceDto? = dbQuery {
+        principal.requirePermission("get device by id", XPrincipal::allowedToGetDevice)
         Device.findById(id)?.let { toDeviceDto(it) }
     }
 
-    suspend fun getDeviceByMac(mac: String): DeviceDto? = dbQuery {
+    suspend fun getDeviceByMac(mac: String, principal: XPrincipal): DeviceDto? = dbQuery {
+        principal.requirePermission("get device by mac", XPrincipal::allowedToGetDevice)
         Device.find { Devices.mac eq mac }.firstOrNull()?.let { toDeviceDto(it) }
     }
 
-    suspend fun createNewDevice(device: NewDeviceDto): DeviceDto = dbQuery {
+    suspend fun createNewDevice(device: NewDeviceDto, principal: XPrincipal): DeviceDto = dbQuery {
+        principal.requirePermission("create new device", XPrincipal::allowedToCreateNewDevice)
+
         val newDevice = Device.new {
             name = device.name
             mac = device.mac
@@ -32,10 +46,19 @@ class DeviceService {
             backupBackendUrl = device.backupBackendUrl
         }
 
-        toDeviceDto(newDevice)
+        val deviceDto = toDeviceDto(newDevice)
+
+        log.domainEvent(
+            "new device: {} by {}",
+            keyValue("deviceDto", deviceDto),
+            keyValue("principal", principal)
+        )
+        deviceDto
     }
 
-    suspend fun editDevice(id: Int, editDevice: EditDeviceDto) = dbQuery {
+    suspend fun editDevice(id: Int, editDevice: EditDeviceDto, principal: XPrincipal) = dbQuery {
+        principal.requirePermission("edit device", XPrincipal::allowedToEditDevice)
+
         val device = Device.findById(id) ?: throw IllegalArgumentException("Device with id $id does not exist")
 
         editDevice.name?.let { device.name = it }
@@ -43,10 +66,22 @@ class DeviceService {
         editDevice.secret?.let { device.secret = it }
         editDevice.bgImageUrl?.let { device.bgImageUrl = it }
         editDevice.backupBackendUrl?.let { device.backupBackendUrl = it }
+
+        log.domainEvent(
+            "edit device: {} by {}",
+            keyValue("deviceDto", toDeviceDto(device)),
+            keyValue("principal", principal)
+        )
     }
 
-    suspend fun deleteDevice(id: Int) = dbQuery {
+    suspend fun deleteDevice(id: Int, principal: XPrincipal) = dbQuery {
+        principal.requirePermission("delete device", XPrincipal::allowedToDeleteDevice)
         val device = Device.findById(id) ?: throw IllegalArgumentException("Device with id $id does not exist")
+        log.domainEvent(
+            "delete device: {} by {}",
+            keyValue("deviceDto", toDeviceDto(device)),
+            keyValue("principal", principal)
+        )
         device.delete()
     }
 
@@ -62,23 +97,41 @@ class DeviceService {
         )
     }
 
-    suspend fun checkDeviceCredentials(mac: String, secret: String): Boolean = dbQuery {
-        val device = Device.find {
+    suspend fun checkDeviceCredentials(mac: String, secret: String): DevicePrincipal? {
+        val device = findDevice(mac)
+
+        return if (device != null) {
+            if (device.secret == secret) {
+                DevicePrincipal(mac)
+            } else {
+                null
+            }
+        } else {
+            val newDeviceDto = createNewDevice(
+                NewDeviceDto(
+                    "new device $mac",
+                    mac,
+                    secret,
+                    "",
+                    ""
+                ),
+                NewDevicePrincipal(mac)
+            )
+
+            DevicePrincipal(newDeviceDto.mac)
+        }
+    }
+
+    private suspend fun findDevice(mac: String): Device? = dbQuery {
+        Device.find {
             Devices.mac eq mac
         }.firstOrNull()
+    }
 
-        if (device != null) {
-            device.secret == secret
-        } else {
-            Device.new {
-                this.name = "new device $mac"
-                this.mac = mac
-                this.secret = secret
-                this.bgImageUrl = ""
-                this.backupBackendUrl = ""
-            }
-
-            true
+    private fun XPrincipal.requirePermission(description: String, permission: XPrincipal.() -> Boolean) {
+        if (!this.permission()) {
+            log.info("$$name tried to $description")
+            throw AuthorizationException("$name not allowed to $description.")
         }
     }
 }
