+++ src/service/ToolService.kt
@@ -1,26 +1,37 @@
 package cloud.fabx.service
 
+import cloud.fabx.application.AuthorizationException
+import cloud.fabx.application.XPrincipal
 import cloud.fabx.db.DbHandler.dbQuery
+import cloud.fabx.domainEvent
 import cloud.fabx.dto.EditToolDto
 import cloud.fabx.dto.NewToolDto
 import cloud.fabx.dto.ToolDto
+import cloud.fabx.logger
 import cloud.fabx.model.Device
 import cloud.fabx.model.Qualification
 import cloud.fabx.model.Tool
 import cloud.fabx.qualificationService
+import net.logstash.logback.argument.StructuredArguments
 import org.jetbrains.exposed.sql.SizedCollection
 
 class ToolService {
 
-    suspend fun getAllTools(): List<ToolDto> = dbQuery {
+    private val log = logger()
+
+    suspend fun getAllTools(principal: XPrincipal): List<ToolDto> = dbQuery {
+        principal.requirePermission("get all tools", XPrincipal::allowedToGetAllTools)
         Tool.all().map{ toToolDto(it) }.toCollection(ArrayList())
     }
 
-    suspend fun getToolById(id: Int): ToolDto? = dbQuery {
+    suspend fun getToolById(id: Int, principal: XPrincipal): ToolDto? = dbQuery {
+        principal.requirePermission("get tool by id", XPrincipal::allowedToGetTool)
         Tool.findById(id)?.let { toToolDto(it) }
     }
 
-    suspend fun createNewTool(tool: NewToolDto): ToolDto = dbQuery {
+    suspend fun createNewTool(tool: NewToolDto, principal: XPrincipal): ToolDto = dbQuery {
+        principal.requirePermission("create new tool", XPrincipal::allowedToCreateTool)
+
         val deviceInDb = Device.findById(tool.deviceId) ?: throw IllegalArgumentException("Could not find deviceId.")
 
         val qualificationsInDb: List<Qualification> = tool.qualifications.map {
@@ -42,11 +53,19 @@ class ToolService {
 
             newTool.qualifications = SizedCollection(qualificationsInDb)
 
-            toToolDto(newTool)
+            val toolDto = toToolDto(newTool)
+            log.domainEvent(
+                "new tool: {} by {}",
+                StructuredArguments.keyValue("toolDto", toolDto),
+                StructuredArguments.keyValue("principal", principal)
+            )
+            toolDto
         }
     }
 
-    suspend fun editTool(id: Int, editTool: EditToolDto) = dbQuery {
+    suspend fun editTool(id: Int, editTool: EditToolDto, principal: XPrincipal) = dbQuery {
+        principal.requirePermission("edit tool", XPrincipal::allowedToEditTool)
+
         val tool = Tool.findById(id) ?: throw IllegalArgumentException("Tool with id $id does not exist")
 
         editTool.deviceId?.let {deviceId ->
@@ -68,10 +87,21 @@ class ToolService {
             }
             tool.qualifications = SizedCollection(qualificationsInDb)
         }
+        log.domainEvent(
+            "edit tool: {} by {}",
+            StructuredArguments.keyValue("toolDto", toToolDto(tool)),
+            StructuredArguments.keyValue("principal", principal)
+        )
     }
 
-    suspend fun deleteTool(id: Int) = dbQuery {
-        val tool = Tool.findById(id) ?: throw java.lang.IllegalArgumentException("Tool with id $id does not exist")
+    suspend fun deleteTool(id: Int, principal: XPrincipal) = dbQuery {
+        principal.requirePermission("delete tool", XPrincipal::allowedToDeleteTool)
+        val tool = Tool.findById(id) ?: throw IllegalArgumentException("Tool with id $id does not exist")
+        log.domainEvent(
+            "delete tool: {} by {}",
+            StructuredArguments.keyValue("toolDto", toToolDto(tool)),
+            StructuredArguments.keyValue("principal", principal)
+        )
         tool.delete()
     }
 
@@ -86,5 +116,12 @@ class ToolService {
             tool.wikiLink,
             tool.qualifications.map { qualificationService.toQualificationDto(it) }
         )
+    }
+
+    private fun XPrincipal.requirePermission(description: String, permission: XPrincipal.() -> Boolean) {
+        if (!this.permission()) {
+            log.info("$name tried to $description")
+            throw AuthorizationException("$name not allowed to $description.")
+        }
     }
 }
