+++ src/service/UserService.kt
@@ -1,24 +1,35 @@
 package cloud.fabx.service
 
+import cloud.fabx.application.AuthorizationException
+import cloud.fabx.application.XPrincipal
 import cloud.fabx.db.DbHandler.dbQuery
+import cloud.fabx.domainEvent
 import cloud.fabx.dto.EditUserDto
 import cloud.fabx.dto.NewUserDto
 import cloud.fabx.dto.UserDto
+import cloud.fabx.logger
 import cloud.fabx.model.User
+import net.logstash.logback.argument.StructuredArguments
 
 class UserService {
 
+    private val log = logger()
+
     private val qualificationService = QualificationService()
 
-    suspend fun getAllUsers(): List<UserDto> = dbQuery {
+    suspend fun getAllUsers(principal: XPrincipal): List<UserDto> = dbQuery {
+        principal.requirePermission("get all users", XPrincipal::allowedToGetAllUsers)
         User.all().map{ toUserDto(it) }.toCollection(ArrayList())
     }
 
-    suspend fun getUserById(id: Int): UserDto? = dbQuery {
+    suspend fun getUserById(id: Int, principal: XPrincipal): UserDto? = dbQuery {
+        principal.requirePermission("get user by id", XPrincipal::allowedToGetUser)
         User.findById(id)?.let { toUserDto(it) }
     }
 
-    suspend fun createNewUser(user: NewUserDto): UserDto = dbQuery {
+    suspend fun createNewUser(user: NewUserDto, principal: XPrincipal): UserDto = dbQuery {
+        principal.requirePermission("create new user", XPrincipal::allowedToCreateNewUser)
+
         val newUser = User.new {
             firstName = user.firstName
             lastName = user.lastName
@@ -28,10 +39,18 @@ class UserService {
             lockedReason = ""
         }
 
-        toUserDto(newUser)
+        val userDto = toUserDto(newUser)
+        log.domainEvent(
+            "new user: {} by {}",
+            StructuredArguments.keyValue("userDto", userDto),
+            StructuredArguments.keyValue("principal", principal)
+        )
+        userDto
     }
 
-    suspend fun editUser(id: Int, editUser: EditUserDto) = dbQuery {
+    suspend fun editUser(id: Int, editUser: EditUserDto, principal: XPrincipal) = dbQuery {
+        principal.requirePermission("edit user", XPrincipal::allowedToEditUser)
+
         val user = User.findById(id) ?: throw IllegalArgumentException("User with id $id does not exist")
 
         editUser.firstName?.let { user.firstName = it }
@@ -42,10 +61,22 @@ class UserService {
         editUser.lockedReason?.let { user.lockedReason = it }
         editUser.cardId?.let { user.cardId = it }
         editUser.cardSecret?.let { user.cardSecret = it }
+
+        log.domainEvent(
+            "edit user: {} by {}",
+            StructuredArguments.keyValue("userDto", toUserDto(user)),
+            StructuredArguments.keyValue("principal", principal)
+        )
     }
 
-    suspend fun deleteUser(id: Int) = dbQuery {
+    suspend fun deleteUser(id: Int, principal: XPrincipal) = dbQuery {
+        principal.requirePermission("delete user", XPrincipal::allowedToDeleteUser)
         val user = User.findById(id) ?: throw IllegalArgumentException("User with id $id does not exist")
+        log.domainEvent(
+            "delete user: {} by {}",
+            StructuredArguments.keyValue("userDto", toUserDto(user)),
+            StructuredArguments.keyValue("principal", principal)
+        )
         user.delete()
     }
 
@@ -62,5 +93,12 @@ class UserService {
             user.cardSecret,
             user.qualifications.map { qualificationService.toQualificationDto(it) }.toCollection(ArrayList())
         )
+    }
+
+    private fun XPrincipal.requirePermission(description: String, permission: XPrincipal.() -> Boolean) {
+        if (!this.permission()) {
+            log.info("$name tried to $description")
+            throw AuthorizationException("$name not allowed to $description.")
+        }
     }
 }
